Fix allocation warning

+++ src/p7_search_builder.c
@@ -152,7 +152,7 @@ p7_search_builder(const ESL_GETOPTS *go, ESL_ALPHABET *abc, char *infile, char *
 #endif
 
   infocnt = (ncpus == 0) ? 1 : ncpus;
-  ESL_ALLOC(info, sizeof(*info) * infocnt);
+  ESL_ALLOC(info, (ptrdiff_t) sizeof(*info) * infocnt);
 
   for (i = 0; i < infocnt; ++i)
   {
